requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
chromadb
sentence-transformers
fastapi
uvicorn
pydantic
requests
pypdf
langchain-community
langchain-text-splitters
google-genai
